+++ resources/lang/en/validation.php
@@ -1,11 +1,11 @@
 <?php
 
 return [
-    'unique'  => 'Value already exists',
-    'required' => 'Field can\'t be blank',
+    'unique'  => 'Invalid email or password',
+    'required' => 'The field is required',
     'regex' => 'Field is not in the correct form',
     'confirmed' => 'Field doesn\'t match',
-    'date' => 'Date is wrong',
+    'date' => 'The date is wrong',
     'max' => [
         'string' => 'Field can\'t exceed :max characters'
     ]