+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+### 0.1.0: 2025-01-02
+
+* Add Laravel 10 support
+* Fix mail configuration
+* Fix missing migrations
+* Fix missing ebillprovider field
+* Fix errors in sign up
+* Obfuscate user exists message for security
+* Fix a typo in .env that would cause the app to crash
+
 ### 0.0.8: 2025-01-02
 
 * Release initial version